services/conversation/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
prometheus-client==0.19.0
structlog==23.2.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
psycopg2-binary
sqlalchemy
kafka-python
openai
groq
google-generativeai
